Fix a sensor problem and a problem with socklen in the LAN code.
git-svn-id: file:///home/cminyard/t2/svn/openipmi/trunk@1865 a860ddb0-99b4-4eb6-8282-105f033f2ff5

+++ lib/sensor.c
@@ -4763,7 +4763,7 @@ typedef struct reading_get_info_s
     void                       *cb_data;
     ipmi_states_t              states;
     enum ipmi_value_present_e  value_present;
-    double                     raw_val;
+    unsigned int               raw_val;
     double                     cooked_val;
 } reading_get_info_t;
